refactor: coupon 관련 utils 함수 위치 변경
